requirements.txt
# Copyright (C) 2023 Anthony Harrison
# SPDX-License-Identifier: Apache-2.0
bids-analyser
lib4sbom